Renamed auth module to users
