+++ backends/api/src/app.ts
@@ -1,7 +1,7 @@
 import express from "express";
 import cors from "cors";
 import passport from "./configs/passport.ts";
-import authRouter from "./auth/controller.ts";
+import usersRouter from "./users/controller.ts";
 import logbookRouter from "./logbooks/controller.ts";
 import requireAuth from "./middleware/requireAuth.ts";
 import morganMiddleware from "./middleware/morgan.ts";
@@ -13,7 +13,7 @@ app.use(express.json());
 app.use(passport.initialize());
 app.use(morganMiddleware);
 
-app.use("/auth", authRouter);
+app.use("/users", usersRouter);
 app.use("/logbooks", requireAuth, logbookRouter);
 
 export default app;
