+++ backends/api/src/users/service.ts
@@ -1,9 +1,9 @@
 import * as repository from "./repository.ts";
-import type { User } from "./../types/user.ts";
+import type { User } from "../types/user.ts";
 import * as auth from "../utils/authUtils.ts";
 import emailValidator from "./validators/emailValidator.ts";
 import { ENV } from "../configs/env.ts";
-import type { Result } from "./../types/result.ts";
+import type { Result } from "../types/result.ts";
 
 export const createUser = async (
   input: Pick<User, "username" | "email"> & { password: string },
